refactor: rename BuilderStepsGeneration folder and namespace
