+++ src/M31.FluentApi.Generator/CodeGeneration/CodeBoardElements/CodeBoard.cs
@@ -1,5 +1,4 @@
 using M31.FluentApi.Generator.CodeBuilding;
-using M31.FluentApi.Generator.CodeGeneration.CodeBoardActors.BuilderStepsGeneration;
 using M31.FluentApi.Generator.CodeGeneration.CodeBoardActors.Commons;
 using M31.FluentApi.Generator.CodeGeneration.CodeBoardActors.MethodCreation.Forks;
 using M31.FluentApi.Generator.SourceAnalyzers;
