+++ src/M31.FluentApi.Generator/CodeGeneration/CodeGenerator.cs
@@ -1,5 +1,5 @@
 using M31.FluentApi.Generator.CodeGeneration.CodeBoardActors;
-using M31.FluentApi.Generator.CodeGeneration.CodeBoardActors.BuilderStepsGeneration;
+using M31.FluentApi.Generator.CodeGeneration.CodeBoardActors.BuilderMethodsGeneration;
 using M31.FluentApi.Generator.CodeGeneration.CodeBoardActors.DuplicateMethodsChecking;
 using M31.FluentApi.Generator.CodeGeneration.CodeBoardActors.InnerBodyGeneration;
 using M31.FluentApi.Generator.CodeGeneration.CodeBoardActors.MethodCreation.Forks;
